+++ src/frontend/src/components/Map.js
@@ -77,7 +77,7 @@ const PredictionPolygon = ({ hullPoints }) => {
     hullPoints && (
       <Polygon
         positions={hullPoints}
-        pathOptions={{ color: "blue", fillColor: "blue" }}
+        pathOptions={{ color: "#4eaee4", fillColor: "#4eaee4" }} // blue
       />
     )
   );
@@ -104,13 +104,13 @@ const PredictionHexagons = ({ predictionHexagonIDs }) => {
     predictionHexagons && (
       <Polygon
         positions={predictionHexagons}
-        pathOptions={{ color: "blue", fillColor: "blue" }}
+        pathOptions={{ color: "#4eaee4", fillColor: "#4eaee4" }} // blue
       />
     )
   );
 };
 
-const AnnotationHexagonsLayer = ({ annotationHexagonIDs }) => {
+const AnnotationHexagonsLayer = ({ annotationHexagonIDs, color }) => {
   const [annotationHexagons, setAnnotationHexagons] = useState([]);
 
   useEffect(() => {
@@ -124,7 +124,7 @@ const AnnotationHexagonsLayer = ({ annotationHexagonIDs }) => {
     annotationHexagons && (
       <Polygon
         positions={annotationHexagons}
-        pathOptions={{ color: "green", fillColor: "green" }}
+        pathOptions={{ color: color, fillColor: color }}
       />
     )
   );
@@ -145,7 +145,9 @@ const Map = ({
   annotationHexagonIDs,
   onAddAnnotationHexagonIDs,
   hexResolution,
+  taxonId,
 }) => {
+  console.log('Render map');
   return (
     <MapContainer
       center={[39, 34]}
@@ -187,18 +189,34 @@ const Map = ({
           <HexagonLayer hexResolution={hexResolution} />
         </LayersControl.Overlay>
 
+        {/* Add the iNaturalist Observations Layer */}
+        <LayersControl.Overlay checked name="iNaturalist Observations">
+          <TileLayer
+            url={`https://tiles.inaturalist.org/v1/grid/{z}/{x}/{y}.png?taxon_id=${taxonId}`}
+          />
+        </LayersControl.Overlay>
+
         {/* Custom Hexagons Layer */}
-        <LayersControl.Overlay checked name="Current Annotation">
+        <LayersControl.Overlay checked name="Annotation (Presence)">
           <AnnotationHexagonsLayer
-            annotationHexagonIDs={annotationHexagonIDs}
+            annotationHexagonIDs={annotationHexagonIDs.presence}
+            color={"#00b175"}
+          />
+        </LayersControl.Overlay>
+
+        <LayersControl.Overlay checked name="Annotation (Absence)">
+          <AnnotationHexagonsLayer
+            annotationHexagonIDs={annotationHexagonIDs.absence}
+            color={"#e14b23"}
           />
         </LayersControl.Overlay>
 
         {/* Render the PredictionHexagons if hexagons are available */}
-        <LayersControl.Overlay name="Prediction Hexagons">
-          <PredictionHexagons predictionHexagonIDs={predictionHexagonIDs} />
-        </LayersControl.Overlay>
-
+        {predictionHexagonIDs && (
+          <LayersControl.Overlay name="Prediction Hexagons">
+            <PredictionHexagons predictionHexagonIDs={predictionHexagonIDs} />
+          </LayersControl.Overlay>
+        )}
         {/* Render the PredictionPolygon if hullPoints are available */}
         {hullPoints && (
           <LayersControl.Overlay name="Prediction Polygon">